[59539] Keep manual scheduling mode in migration
The rule is: it never switches from manual to automatic scheduling mode.
It only switches from automatic to manual, and only if keeping automatic
is not possible because it would mean losing the dates.

The code and specs have been updated to reflect this.

A materialized view is used to reuse the data in multiple different
queries.