[59539] Reschedule when switching parent to automatic mode
When a parent work package switches from manual to automatic scheduling
mode, the children dates need to be known to be able to set parent
dates.

So a scheduling is done first in the `SetAttributesService` to get
children's dates so that parent dates can be set. Then the parent work
package is saved (in `UpdateService`), and all dependent work packages
are rescheduled (again).