Preload DatePicker PreviewController to make tests pass
When in tests the date picker is opened to click on a day of the
flatpickr, the controller is not loaded yet and some events are missed,
making the test fail.

The controller is not ready yet because it needs to be loaded
dynamically by the OpApplicationController. By registering it in the
setup.ts, it is loaded before the date picker modal is opened, and the
controller is connected before a day is clicked, making (hopefully) the
test pass.