Improve month selection for date picker tests
After selecting a month, `#select_month` returns the index of the
selected month. It previously worked well only for single-month
datepickers. For a multiple months date picker, it was failing because
the selector is invalid (so it waited 5 seconds for nothing), then it
was retried, the current month was calculated, and that was the value
returned as the 2 conditions were not met. So in the end it was
returning the correct value, but it was also doing one useless retry.

This commit introduces `#current_month_index` to return the index of the
current month. This one works for both single and multi date pickers. It
is used by `#select_month` to calculate the number of months to move
forward or backward.

This should lead to some faster tests.